feat: add pending_application_resource link table
Will be used by the AddResourcesBeforeApplication to link a resource
with an application name, where the application does not exist yet.
These rows should only exist for during the deployment process and be
removed when the application is created in the juju model. At that time
the resources should be linked by the application's UUID.