feat: add domain methods related to pending resource concept
AddResourcesBeforeApplication and RevisionResource are new resource
domain methods to break apart the overly complex AddPendingResources
resouce facade call into logical groupings.

AddResourcesBeforeApplication allows a resource to be added after the
charm is known to the juju model and before the application exists. This
will be resolved into a regular resource when the application is created
later. Used for deploying a bundle.

UpdateRevisionResource updates the resource to a store resource origin and
the given revision number. The application's charm modified version is
also incremented to trigger a charm refresh. Remove stored resource if one
exists otherwise the resource opener will find the blob and use the old one
rather than download the new.

Used when refreshing an application's resource to a new revision.

DeleteResourcesAddedBeforeApplication will take the given resource UUIDs
and delete the related rows from the resource and
pending_application_resource tables.

This method is required for deployment scenarios which cannot be
accomplished in one server side transaction, e.g. bundles and local
charms, where the deploy may fail part way thru the process.