feat: exposes objects handler
Exposes new handler to get an object (charm, resource, agent binary)
from the object store. It doesn't matter what it is, we'll get it
by a certain hash and return it. This is going to be useful for
enabling HA file based object store. If another controller node
want's that sha, then it just asks other nodes in the cluster if
they have it, otherwise return NotFound.

There is no ratelimiting on any of these end points, there really
should be, it's exposing the guts to everyone.