feat: wire application constraints in the apiserver
This patch wires the recently added application constraints methods from
the application domain into the apiserver.

As a side quest, the `Constraints()` method from the legacy state
`Application` struct was made private, since it's only used inside the
legacy `state/application.go` and making it private we ensure it's not
used elsewhere. It must be noted that we cannot remove it because
setting the constraints in mongo has to continue until the machine
domain is finished.