feat: non-tracked provider
This introduces the concept of the non-tracked provider. The
provider will be a provider that has a life-cycle and will be
associated with a runner, but when it's done with, we will remove
it from the runner. This is useful for when we need or want a
provider, but we can't look at the database for the relavant config
and information. This comes up during the creation of a model, or
when importing it. In addition to this, the non-tracked provider
will not handle credential invalidation, this will fall in line
with the new credential invalidator work[1]

  1. https://github.com/juju/juju/pull/18840