feat: resolve pending resources in create application
Update CreateApplication to handle resources created before
the application via the resource domain, as well as continue
to create resources that were not previously created.

Local charms and bundles will add resources after the charm
is set and before calling Deploy. For charmhub charms, the
client will do this all in one call. The two scenarios can
not be mixed.

AddResourcesBeforeApplication will not add potential resources
so do that when moving them from pending to real resources.
The potential resources are not added so that only actual
resource UUIDs are passed to and from the juju client which
helps ensure compatibility with older clients.