refactor: remove the runner
Rename the non-tracked worker to just be an ephemeral provider.

People believe that we don't want to keep ties on the ephemeral
providers, which I can see the point of, but we do loose the
modelling from an application sense. We now no longer can visibly
see how many epemeral providers are currently being used, and how
long they're being used for.