fix: stop handling unit and placeholder resources in migration handler
The resource migration HTTP handler recieves resources from the source
controller during migration and sets them in the target controller. In
previous versions of Juju, this migration upload handler was also used
to transfer metadata about the resources. This is no longer needed in
4.0 as the metadata is transferred during the database migration step.

The metadata transferred was specifically about unit resources, and
placeholder resources (signaled by a timestamp of 0). The unit resources
indicated which unit was using which resource, and the placeholder
resources were resources that did not yet have blobs uploaded to the
controller.

The resource migration handler was one of the first things to be touched
during the resource epic, and was done before the rest of the resource
epic. It mistakenly handled unit resources as well as application
resources. It would call SetUnitResource on unit resources, which would
mark them as in use by the unit. This is information that would already
be in the target controller as it is set as part of the database
migration.

Placeholder resources were not handled, as they should be.

In this change, uploading a unit resource or a placeholder resource will
cause the handler to return status OK, but take no action.

Previously, the handler would return information about the resource that
hand been uploaded. This is no longer done for placeholder and unit
resources. Normally I would err on the side of caution and continue to
return such information, but since this HTTP handler is expliciatally
for migration, we know that the only client who will be using this is
a Juju controller of version 3.6 or above, since nobody else can carry
out migrations. In 3.6 (and above), the returned data from this handler
is dropped.