chore: stop uploading unit resources and remove related client methods
In 4.0 and above there is no need to migrate unit or placeholder
resources via the HTTP handler since all this information is migrated as
part of the database migration.

In this change, we stop uploading unit resources and remove the client
methods for uploading unit and placeholder resources (the placeholder
upload functionality was already unused).