fix: use after release
The cloud validate credential was releasing the model back to the
state pool, before using it. This is a classic use after free. This
manifested itself when running API tests in JIMM against the Juju
API server and the test would panic, stating the session was already
closed.
Juju is much more stricter when we close a session[1], to prevent
leaks. All sessions will close on state pool Release or Remove when
the reference count is 0. Thus, attempting to access a model after
it has been released is no longer available, because the session
has already been closed, as there are no references to it.

I suspect this won't be the last of these though.

 1. https://github.com/juju/juju/pull/18614