fix: allow db repl to switch to models containing "-"
The ".switch" command in the DB repl worker used to split by "-" and
check that there were two parts. This meant would declare the model name
invalid if it contained a "-" as it only expected to find one, located
in the prefix "model-".

Change to check for the specific prefix rather than splitting the string
by "-".