chore: move service interfaces close to usage
This whole piece of work started because I wanted to break up the
`apiserver/common` package. This package had slowly accumulated many
disparate entities, so had little coherence.

This became a problem when different aspects of the package use the same
service, but different parts. Interfaces clashes, or had to be expanded
artificially to include the union of both use cases.

Lack of coherence within a package obviously points to the need to break
up the package, which is exactly what I then did.

I created _a single subpackage_, `model`, which contains all the 'common'
apiserver operations to do with models. This itself was very easy.

However, I then discovered that maybe two dozen facade packages import
service interfaces from common, interfaces which had now been moved and
split.

So instead of maintain the problem and simply rename the imports, I have
decoupled the facade packages (to the greatest extent possible) from
common by moving these interfaces to where they are used. This is the
pattern we have mostly been following, but there are many exceptions.