refactor: update application exposed endpoint
The application exposed tables for both cidr and space didn't have
foreign keys to the relation endpoint tables. Strictly adding those
allows us to see which endpoint was related to expose call.

The relation endpoint uuid can be NULL, which means that the exposed
space or cidr is for the wildcard endpoint; for all application
endpoints.

It's possible for an application to have multiple spaces and cidrs
associated with the same relation endpoint.

The representation then becomes:

app id | relation endpoint id | space id
----------------------------------------
1      | a                    | s1
1      | a                    | s2
1      | b                    | s1
1      | b                    | s3

app id | relation endpoint id | cidr
-------------------------------------------
1      | a                    | 10.0.0.0/24
1      | b                    | 0.0.0.0/0
1      | b                    | ::/0
1      | NULL                 | 0.0.0.0/0
1      | NULL                 | ::/0

To get all the exposed endpoints for a relation a view is created
to UNION both tables.

The UNION of both tables would then become:

app id | relation endpoint id
-----------------------------
1      | a
1      | b
1      | NULL

So the application (1) has three exposed endpoints:

 1. The wildcard
 2. To a
 3. To b

This replicates what we have with mongo.