chore: re-write get & set status apiserver common tests
A number of tests in the apiserver/common package dealing with setting
and getting the statuses of entities were being skipped, because they
were based on the legacy state factories and thus were failing with the
transition to DQLite.

Re-write these tests with mocks, and re-enable them. This included some
subtle additions to the production code itself to make it more testable,
such as injecting a clock instead of depending on time.Now()