feat: set application status for unit leader
Add a method for setting the application status for only the unit
leader. The application set status common facade method ensures
that you can only set the status for the unit leader.

This threads in the leadership ensure into the application service
along with adding a noop leadership ensurer for testing purposes.