feat(facade): expose the api clock on the facade context
In many api facades we were using the WallClock when a clock was
required for operations.

However, the apiserver is granted a clock by a worker, which should have
been used instead. To make use of this, we need to expose it on the
facade context.

Add and implement a new method to the facade context interface. Wire up
this new clock everywhere we were using a WallClock or a clock from
SystemState.

This means the clock in state is kept private within state. Remove the
SystemState .Clock method.