feat: wire up AddPendingResources to use service calls
Use the new resource service methods which disambiguate
AddPendingResources by breaking apart the functionality. Either the
client wants to add resource info for deploying an application
before the application exists. Or the client is updating the revision of
the resource by refreshing the application.

Add a doc file to explain AddPendingResources more in depth.