feat: wire up provider credential invaldiator
The following is the wire up of the provider credential invalidator
from the environs package. It is currently a no-op and it breaks
all providers because the worker hasn't wired this up yet, but when
it does this is the right way.

The providers are a **mess** and I mean a mess when it comes to this.
There isn't one consistent way of doing this. There should be a
common base credential provider that each one can compose from. Then
you can construct it with a auth failure predicate and the credential
invalidator. Then you can just call a method on the provider with
a context and an error.

The providers lack vision and design and it's easy to see.