chore: remove unused information from migrationmaster Export
Remove the unit and charm resource information from the migration master
APIs Export method and rev the facade.

The information about the unit and charm resources is no longer needed
in the migrationmaster worker due to a previous change. The information
was still transmitted in the facade.

Not transmitting this information will allow for easier changes to the
description package as it is fetched from the exported model
description.

The facade is only used by the controller to communicate with the
migrationmaster worker. Therefore, revving the facade should not cause
any issues with migration to and from other clients.