feat(resources): add update resource
This update resource function adds a new row to the resources table and
updates the application_resource table to point to the new resource. It
then increments the charm modified version.

It was realised that an issue in the resource domain meant that the
revisions of unit resources and kubernetes application resources were
not being correctly tracked. This was becuase these tables only linked
to the resource table, and did not themsevles keep track of the revision
or origin. The resource table was being updated when a new revision was
added thus implicitly also updating the unit and kubernetes application
resources versions.

The resource schema was designed so that when a new resource revision
was added, a new row would be added to the resource table. When a blob
was uploaded it would be linked to this new resource. The old resource
record remains, and is still linked by the unit and application, which
will update to the new version of the resource after their watcher spots
the change.

There will need to be some worker that removes old orphaned resources.
As discuessed with @Joe, this is the cleanest way to deal with these
old blobs and old resource records.