feat(status): wire up unit status domain methods to common
At the same time, refactor the common status mechanisms.

The original implementation of status getter and setting in common uses
a number of bad abstractions. We force into the same place our
mechanism to set the status for any entity. This means that, in theory,
our machine api could set the status of an application, unit, volume,
etc. if it is misused.

Create a new series of structs in common designed to only manipulate the
status of units. Wire this new struct up, and reject unit tags on the
existing common api methods.

The new structs are backed with DQLite. However, we read the
status of units in other places without going through the common
package, so for the time being we need to dual-write to mongo.