refactor: update resource to be specific to uploaded resources
UpdateResource becomes UpdateUploadResource. The method only works for
resources which will be uploaded after the application has been
deployed. This is the corresponding method to UpdateResourceRevision.
Additionally it will now delete the current resource blob if one exists.

The charm modified version is not updated in this direct code path, but
rather will be when the upload actually happens.