fix: update resource before saving upload
If a upload request includes a pendingID, that is the updated resource
UUID, use it. If there is no pendingID, call UpdateUploadResource before
storing the blob to prepare juju for the new blob.

Deploy and Refresh scenarios from the juju cli, both call
AddPendingResource first, which take care preparing for the new blob.
However juju attach-resource does not. This update takes care of that
for compatibility purposes.