feat: introduce is controller model flag
This commit introduces a new is_controller_model flag to the v_model
view. This allows the user to see at a quick glance if this model is for
the controller. This change was brought about to support list summaries
of models. The current implemnentation we had doesn't work properly when
no controller model exists and this is a bug. To fix this we are taking
the foot guns out of some of these util functions.