fix: lxd VM instance creation due to duplicate alias
When we try to create a VM instance on lxd (e.g `juju add-machine
--constraints="virt-type=virtual-machine"`) we'll get an error stating
that it cannot start because of a duplicate alias. This happens because
when we copy the lxd images to the local cache we copy the aliases as
well, and the original alias is the same for both containers and VMs
(e.g. "24.04/amd64").
This patch fixes this behavior by simply removing the original alias
from the local image cache and therefore we are able to copy the VM
image.