refactor(status): use domain instead of state for application status
Make use of the newly modified GetApplicationStatus to retrieve the
status from DQLite instead of Mongo.

You may be wondering why this is included in the task to wire up *unit*
statuses. This is because our facade handlers would also query the unit
statuses and use those to derive the application status if it is unset.

However, this logic has now been encapsulated into the domain, so we can
drop all of this.

As a result, we can also drop a lot of code out of core/status.

NOTE: We should also take into account the container status CAAS models.
We do not do that yet in the domain. However, that will be done next so
we don't need to worry about it atm