feat(status): add DisplayStatus methods for apps and units
We (kinda) have a concept of a display status for applications and
units. This status reduces to the workload status for IAAS models, but
for CAAS models takes the 'container status' and 'operator status' into
account when calculating the status for units and applications
respectively.

The business logic for how to do this was contained within core/status.
Copy and convert this logic over into the application domain state
layer.

Add two service methods using this, backed by stub state.
GetUnitDisplayStatus, and GetApplicationDisplayStatus.

NOTE: It is impossible to cleanly move this logic over, since we have
stronger type constraints on workload & cloud container statuses.