feat: wind out the agent from modelworkermanager
We don't need to double write the logsink twice when we start
the model worker manager. The logsink is the logger context. This
means we just need to request that and it has everything already
setup.

This means we can wind out the agent from the modelworkermanager
and just use the id from the agent in the logsink and not in the
modelworkermanager.

We've got the same functionality, but nicely encapsulated inside
of the logsink. Later we'll push the logsink into the domainservices
and all this refactoring will become a lot cleaner.