fix: add resource row in UpdateResourceRevision
Update the UpdateResourceRevision method to add a new resource row and
link it in the application_resource table. The same behavior as
UpdateUploadResource method does.

It will also return the new resource UUID for use by AddPendingResource.
AddPendingResource in the resource api package requires the facade call
to return a UUID for each resource given.