feat: thread logsink into domain services
This is it! This threads the logsink into the domain services. This
is the work required for status history. With this we should be able
to get all of status history working nicely.

In order to get a logger context we need to use a LoggerKey which
is REALLY annoying as we have to do a DB request to get the information.
To prevent lots of requests to the readonly table there is a weak
pointer cache to hold the contexts, although it's probably better
to hold the keys (I'll change that!).