chore: move jwt parser into a new worker
The JWT authenticator in apiserver/authentication/jwt creates a JWTAuthenticator which is responsible for decoding JWTs from JIMM and then using these JWTs for authentication and authorization in the api server.

With the new sshserver, there is a new entrypoint for clients that will also require authz/authn. The sshserver also requires a JWT parser which it can then use to make auth decisions.

For this reason, the JWT parsing logic has been moved into a separate worker so that it can be passed as a dependency into both the apiserver and sshserver workers.