feat(resource): stop updating origin and revision through StoreResources
Since the new way to use resources tables is to add a resource and move association through
application_resource and unit_resource table, StoreResources doesn't need to update origin or
revision anymore. It will always upload a resource with valid origin and revision.

So this commit remove those fields from all layers in the call stack.

The only caveat is in resourcemigration handler, we check the compliance between query parameters
and actual stored resources regarding origin and revision to avoid storing the wrong data for the
wrong resource in DB (the UUID is queried from application and name in the handler and can be
subject of data race)