fix(status): fix GetUnitWorkloadStatus to return the correct errors
It turns out we have previously missed a test case for
GetUnitWorkloadStatus when the unit doesn't exist, meaning the incorrect
error was returned.

Fix this bug.

Do this by adding a checkUnitExists method, equivalent to the
pre-existing checkApplicationExists. To match this function, return a
UnitIsDead error if the unit exists, but is dead. This will likely be
useful in future

As a flyby, move some basic query structs inline