refactor: update controller state
This change refactors controller state model summary methods to be safer
and more pragmatic about what they need to achieve. We know longer error
when a controller record doesn't exist in the database. This is on
purpose as the main function of listing model's can still be achieved.

We also simplified the SQL to better answer the questions of list
model's.