refactor: controller state for controller model
This commit introduces a new method required by the service layer for
fetching just the UUID of the controller model if it exists. Because of
this new function we are not going to support lookup of model
information based on name and owner and instead will require a uuid.

This commit refactors model cloud and credential function to perform
this lookup based on UUID and not model name as per above.