fix(state/charm.go): fix for AddCharmMetadata buildTxn
The method was checking if the charm doc existed and if it didn't it would call insertCharmOp, which
again checked for the doc existence. This meant that if somebody tried adding a charm multiple times
in parallel, this would error out. Now we try to create insert charm ops first and, should that fail,
fall back to update.

fix #2071599