feat(statuses): add domain method to replace UnitStatuses on state
This method used to return the statuses for all the units on a
particular application.

Fortunately, we already have a state method which does exactly that!

However, it needed to be modified slightly, since the api indexed
statuses by unit name, whereas our state method did so via unit UUIDs.
Patch this method to follow suit