feat(relations): wire up dummy relations service to application facade
Wire up the NotImplemented relations service to the client application
facade.

The AddRelation facade is changed to only support adding simple Relation.
Support for Peer relation has been dropped, since it will be done in application
domain.
The InferEndpoints method is not replaced in the domain, instead, the
logic it contains will be moved to the service layer in AddRelation.

relationData used in UnitsInfo has been replaced by a direct call to the domain.
crossModelRelationData, once used in relationData, has simply been removed
because CMR will be implemented in their own domain.

SetRelationsSuspended return an error as not implemented,
since this call is only supported for CMR, which will be
implemented in their own domain.

The body of the DestroyRelation facade method is replaced with
NotImplemented, as this work will be done at a later data.