refactor(status): use domain for application display status
Make use of the new GetApplicationDisplayStatus service method to
calculate the application display status, backed by DQLite.

You may be wondering why this is included in the task to wire up *unit*
statuses. This is because our facade handlers would also query the unit
statuses and use those to derive the application status if it is unset.

However, this logic has now been encapsulated into the domain, so we can
drop all of this.

As a result, we can also drop a lot of code out of core/status, and
(*Application).UnitStatuses from the state package.

Also, as a flyby, restore apiserver/facades/agent/uniter/status_test.go
tests. ApplicationStatus, the method being tested, no longer depends on
Mongo state, so we can remove the factory methods and move over to
mocked services entirely.