refactor: split CreateModel
CreateModel, which was modified to accept an agent version now becomes
CreateModelForVersion. CreateModel now calls this method passing the
default agent version.

This was done because only model import needs to specify the version
expjlicitly. Normal model creation is done with a default version, which
is now determined in the service layer instead of passing the current
version from the API.