feat: azure environs credential invalidator
Follows on from [1] and breaks pieces off from [2].

Adds the credential invalidator to the provider directly, so it's
possible invalidate a credential without passing the invalidator
directly onto the call context.

The whole premise behind this is to not need to pass the credential
service into other services. The provider tracker will do all of
that work for us.

 1. https://github.com/juju/juju/pull/18936
 2. https://github.com/juju/juju/pull/18919