chore: ensure that space constraints are decoded even if missing exclude
Before this commit, the space constraint was only decoded from it's DB
value if both the space and the exclude values are valid (i.e. not
NULL). This works fine, but it's better to be a little defensive and
always decode them even if the exclude value is not set, in that case we
assume it's a positive constraint (not excluded).