feat(status): cut over ModelPresenceContext to use domain
The ModelPresenceContext struct, which exposes some methods which mutate
statuses to take into account the 'presence' of entities, is now backed
by DQLite.

The method UnitStatus, however, in different places would work upon
either the workload status or the display status, hidden behind an
interface.

This suggests that the abstraction is wrong. So I have refactored this
method to accept the unit status and agent status in it's params, and to
leave it to the caller to decide where to get these statuses from.

NOTE: the client status facade would read the unit status from a cache,
instead of Mongo directly. I have dropped this caching layer. This is
because, in the long term, we will be re-writing status to use it's own
domain.