refactor: aws credential invalidator
Following on from[1][2], this continues on the work of removing
the provider credential invalidator from the providers. This is
the AWS EC2 one. This uses the new common.CredentialInvalidator
and to invalidate the credentials. The only difference from this
provider over the prior two PR changes, is that we give better
error messages and so we have to wrap the common one with our
own that handles those cases better.

  1. https://github.com/juju/juju/pull/18936
  2. https://github.com/juju/juju/pull/19007