refactor: remove default credential support
This commit removes code that was setting an empty to credential to that
of the controller model when creating a new model. This was a layer
violation and is not needed at the moment. Currently this is safely
handled in the facade and should be more well thought out before
impleneting.