feat(statuses): cutover the uniter facade to use DQLite for statuses
Use the application service instead of Mongo state to find the workload
statuses of applications.

This required dropping a number of factory-based tests. Since this
facade still depends so heavily on Mongo, it is not reasonable to drop
the factories from these tests yet.

So instead, drop the tests and write out a detailed description in
plain English summarising the test coverage, for when we come to restore
these tests