test: re-enable keyupdater watcher tests
Keyupdater watcher tests were disabled because they had flakiness. The
flakiness came from the fact that the test harness needs to consume the
initial event offered by a watcher to work.

However in this case the watcher being tested was a multi watcher that
has n watchers inside of it. This means we have n initial events to
consume.

The fix here is to make the test aware of the fact that there will be n
initial events to consume but it does point out a larger problem with
the multi watcher where the fact that we have a multi watcher as an
implementation detail leaks it's way out to the caller. They now have to
know that there is n events to consume because the watcher is composed
of n other watchers.