feat: add unit constraints as a merge of app and model constraints
This patch persists the constraints resulting from the merge of the
application and model ones, at the time of unit creation.

This replicates the exact same behaviour as we had in the legacy mongodb
state, where we computed the merge before creating a unit.

For the moment this is only verifiable on the db since units have not
been migrated to dqlite yet.