fix: change jaas snap content interface path
Using a content interface and mounting the contents of the jaas snap on top of $SNAP/usr/bin was overwriting the folders contents and preventing the Juju snap from having access to binaries like scp and ssh. In cases where you had the jaas snap installed alongside the juju snap, this  prevented bootstrap no simplestream binary existed, because of a failure to scp the files to the provisoned machine.