fix: stop verifying fingerprint and size for migrated resources
When handling incoming migrated resources, ideally we would verify that
the hash and size of the blob in the request body matches the hash and
size in the headers. However, there is a bug for container resources
exported from 3.6 where the hash the header does not match the hash in
the body. For this reason, this change removes the download size checker
from the migration handler.