fix: handling logout then login for oidc login flows
Previous changes that introduced support for federated login with OIDC/OAuth 2.0 introduced a parameter on `AccountDetails` to indicate whether to use username/password or a session token.
If the user was logged out however, this field didn't exist and username/password auth was always used and failed. This change puts a new value on the `ControllerDetails` struct instead which indicates whether the controller supports OIDC login and defaults to that.