feat: adds context to observer
The observer allows us to hook into the apiserver when a join,
login and leave happens. We want a context.Context, because when
we make a database request, we must have that context to prevent
the database locking up the rest of Juju. With a context.Context,
we can cancel the transaction and dqlite can sort itself out.