refactor: remove presence worker
We don't need the presence worker or the host of pubsub calls to
know if an agent is alive or not. Also, it didn't work correctly
in HA mode[1], as pubsub messages can fail. Instead the plan is
to just write the information into the database. That work will
come later, but for now rip it all out.

---

This is currently broken, but will be fixed.

 1. https://bugs.launchpad.net/juju/+bug/2020849