chore: add a separate provider for supported features
This adds a separate provider for SupportedFeatures on the application
domain.
We need it to be separate because if we have only one provider interface
and only one provider getter, then on substrates other than k8s the
provider will not be supported since the SupportedFeatures method is
only implemented in k8s.
This is an example of what needs to be done for other providers as they
get bigger.