refactor: controller state controller uuid model
This commit refactors the models controller state and how we get the
controller uuid. Previously we were doing this via a seperate statement
but a recent change has added this to the controller view.

We now use the view to get this information. The view has also been
modified to make the controller uuid a left join so models will always
work regardless of the controller.