feat: remove presence
This removes the presence worker and the core package. We no longer
write the presence to the local memory model. All dispatching of
connection topics are dropped on the floor. This will be changed later
when we move to the database holding the presence information when
connecting or disconnecting.

This is a lot of complexity that can just go away in terms of workers.