test(secrets): fix SecretManager facade's TestRemoveSecrets
Tests for the SecretManager facade include a white-box test for removing an entire agent secret.  This test mocks out the internal calls to `secretState.*`, including mocking `secretsState.DeleteSecret`'s return to include a slice of revisions.

DeleteSecret only returns a slice of revisions when it deletes only some of the revisions on a secret, not when it deletes the whole secret.  Thus, as previously written, this test isn't really a test of deleting a whole secret but just deleting some revisions - essentially a duplicate of TestRemoveSecretRevision.

This commit updates TestRemoveSecrets to return from DeleteSecret as if an entire secret was deleted.