fix: use reference name when decoding charm locator
We were incorrectly using the Name in the charm locator struct instead
of the reference name when passing it through to retrieve the charm id.
This patch adds a check to validate that the reference name is not
empty, in that case use it as Name in the charm locator. If it's empty
then use the Name (this case shouldn't happen but it's better to fail
when trying to retrieve the charm).