feat(relation): add 2 methods for uniter facade
Adding GetRelationUUIDByID and GetRelationsStatusForUnit for use by the
uniter facade. RelationsStatus should not be confused with
RelationStatus. The former contain details about the relation, e.g.
Life, that the uniter requires. The latter is able setting the
relation's status for all to use and see, e.g. Suspended.

GetRelationKey updated to use the new type, rather than a string.