fix: temporarily inform the uniter the unit has no relations
Allows uniter to work while relations reimplemented. The uniter
requires two methods to be successful to start up.

One is WatchRelation, required for RemoteStateWatcher to keep working.
The api method must return a minimal watcher which returns an emtpy
string along the channel and that's it. This mimic behavior when a unit
has no relations.

The other is RelationStatus which returns the stutus of all of the
unit's relations. Again mimic behavior when a unit has no relations.

This commit should be reverted once the two methods can successfully get
data from the new relation domain.

This change caused relationResolverSuite tests in the uniter to fail.
Most of the tests are skipped until RelationsStatus can call down into
the domain state layer. To help remember they need to be enabled again,
3 tests have been updated to pass and will start failing once the Skip
should be removed.