feat: wire controller objectstore in domain
This wires up the controller objectstore in the controller damain
services. It will then be available to controller services. At no
point should we be able to access other models within the controller,
to prevent that we don't allow the passing in of a model UUID or
namespace string.

I'm not sold on the name of ModelObjectStoreGetter, as it's no longer
just for models. I'll happily change this, it was previously called
SingularObjectStoreGetter, but was changed to have the model prefix
instead. Considering it is used in both places (models and controller)
I think we should change.