feat: use worker to fix races
There is a race that exhibits in the tests that leaves goroutines
running until they've reached a synchronisation point (ctx.Done or
tomb.Dying) before it will cancel. The solution was to not hand
role our own goroutine management, instead fall back on to worker
runner. This allows us to namespace the worker such that it's
retrivable and identifable when performing the requests. What is
more, because everything is a worker, we can guarantee that all
the retrievers will correctly die and that there are no more
goroutines still processing, esp. when iterating through the tests.