fix: ssh server missing jump host key fix
The server requires the generated jump host key, currently the key is generated and stored in mongo
alongside some other initialisation parameters. We have 3 options, 1) migrate initialise to DQLite
2) Call the mongo state within the worker 3) Generate a key for now to quiet the logs with the
worker bouncing. I've opted for 3) to save time and as 4.0 isn't released right now, I don't think
it is worth the effort to plug mongo into the worker only to remove it later in favour of DQLite.