test: remove mongo tests suites
Removes the mongo test suites. This a bold step. There is no point
spending cycles updating tests that rely on mongo that for intents
and purposes is dead code. The tests may provide some value for test
case behaviour, but that can be looked at in the 3.6 branch or in
the history of this branch.

This releases us from mongo and should provide us the affordance to
move at a faster pace.