feat(status): pull out all status service methods into a separate domain
Create a new domain for the statuses relevant to applications. Move all
methods from the application domain (both for applications and units) to
the new domain.

Cut over the status-related calls on the application service to the new
service.

As a result, we can remove the status-related methods from the
application service.

Next, we will need to move migration code to the new domain package.
Once this is done, we can clean up the state package further.