chore: remove test factory use from uniter tests
Tests embedding uniterBaseSuite were already being skipped, so this
change is non-functional.

We remove state entities and usage of the testing factory from the base
suite and its derivatives altogether. Empty test methods and indicative
comments have been retained for documentation puposes. These should be
deleted as the coverage is assured here or in the domain services.